validPlacement() and Player Constructor created and pass tests.
validPlacement() takes input of where want to place a ship and checks whether one already exists in that spot on the gameboard, if the values are to high/low, and whether already have 5 ships.

If false on all tests will pass the location over to the ship creator, ship hp is location.length.

Player constructor just takes whether player or computer and generates a new gameboard for them.