All parts of receive attack complete, more validations in validPlacement
Taking coordinates for the attack checks whether to many targets,  if it is a miss or not, if it is a hit will reduce hp of the ship if not increment miss on the gameboard. If HP is 0 will mark ship as sunk. If all are sunk game will be over.

Still need to add tracking for what has already been hit so cannot hit same spot twice.