HTML frame built and classed. Attack logic changed.. Random coordinate generator made.
Added when a ship is sunk it's location is changed to false and added the code elsewhere to deal with that.

Random coordinate generator made for computer turns.

Also made 100  HTML buttons for both Player and Computer with their classes to be the nodes of the gameboard for the listeners to be attached to.